implement number spreaders in lobby ctrl
